Check for the user id in book actions

+++ src/controllers/books-controller.js
@@ -3,13 +3,12 @@ const { validate, bookSchema } = require('../utils/validations')
 
 module.exports.create = async (req, res, next) => {
   const error = validate(req.body, bookSchema)
-  if(error) {
-    console.log("Error", error)
-    return res.status(422).json({ error: error.message }) 
-  }
+  if(error) return res.status(422).json({ error: error.message }) 
+  if(req.currentUser._id != req.body.user) return res.status(403).json({ error: 'Unauthorised request' }) 
+
   try {
     const book = await Book.create({
-      user: req.body.userId,
+      user: req.body.user,
       isbn: req.body.isbn,
       title: req.body.title,
       author: req.body.author,
@@ -27,8 +26,9 @@ module.exports.create = async (req, res, next) => {
 }
 
 module.exports.delete = async (req, res, next) => {
+  const userId = req.currentUser._id
   try {
-    const book = await Book.findOneAndDelete({ _id: req.params.id })
+    const book = await Book.findOneAndDelete({ _id: req.params.id, user: userId })
     if(!book) return res.status(404).json({ error: 'Book not found'})
     res.status(200).json({book: book})
   } catch(err) {
@@ -37,8 +37,9 @@ module.exports.delete = async (req, res, next) => {
 }
 
 module.exports.getAll = async (req, res, next) => {
+  const userId = req.currentUser._id
   try {
-    const books = await Book.find()
+    const books = await Book.find({ user: userId} )
     res.status(200).json({ books: books })
   } catch(err) {
     next(err)
@@ -46,10 +47,11 @@ module.exports.getAll = async (req, res, next) => {
 }
 
 module.exports.getOne = async (req, res, next) => {
+  const userId = req.currentUser._id
   try {
-    const book = await Book.findById(req.params.id)
+    const book = await Book.findOne({ _id: req.params.id, user: userId })
     if(!book) return res.status(404).json({ error: 'Book not found'})
-    res.status(200).json({book: book})    
+    res.status(200).json({ book: book })    
   } catch(err) {
     next(err)
   }
@@ -58,6 +60,8 @@ module.exports.getOne = async (req, res, next) => {
 module.exports.update = async (req, res) => {
   const error = validate(req.body, bookSchema)
   if(error) return res.status(422).json({ error: error.message })
+  if(req.currentUser._id != req.body.user) return res.status(403).json({ error: 'Unauthorised request' })
+
   try {
     const book = await Book.findOneAndUpdate({ _id: req.params.id }, req.body, { new: true })
     if(!book) return res.status(404).json({ error: 'Book not found'})
